requirements.txt
Pillow==8.4.0
psutil==5.8.0
PyAutoGUI==0.9.53
requests==2.22.0
